Add files via upload
Uses the sales register of two years (22-24) of a retail company to analyse monthly demand for each SKU. Also, 20 products with the most volatile demand were computed. Applications of this program: With data of sufficient years, we can train a model to predict demand based on the month and SKU and manage inventory better. Limitations of this program: Insufficient data for training. 
Also, this program adjusts the given (untidy) csv files into neater, more legible data frames.